refactor: BaseActivity 네이밍 변경 및 패키지 네이밍 변경
