+++ app/src/main/java/woowacourse/shopping/ui/common/DataBindingActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.base
+package woowacourse.shopping.ui.common
 
 import android.os.Bundle
 import androidx.annotation.LayoutRes
@@ -6,7 +6,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
 
-abstract class BaseActivity<BINDING : ViewDataBinding>(
+abstract class DataBindingActivity<BINDING : ViewDataBinding>(
     @LayoutRes private val layoutResId: Int,
 ) : AppCompatActivity() {
     private var _binding: BINDING? = null
