+++ app/src/main/java/woowacourse/shopping/ui/products/ProductsActivity.kt
@@ -8,13 +8,13 @@ import androidx.activity.viewModels
 import androidx.recyclerview.widget.GridLayoutManager
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityProductsBinding
-import woowacourse.shopping.ui.base.BaseActivity
 import woowacourse.shopping.ui.cart.CartActivity
+import woowacourse.shopping.ui.common.DataBindingActivity
 import woowacourse.shopping.ui.productdetail.ProductDetailActivity
 import woowacourse.shopping.ui.products.ProductsAdapter.OnClickHandler
 
 @SuppressLint("NotifyDataSetChanged")
-class ProductsActivity : BaseActivity<ActivityProductsBinding>(R.layout.activity_products) {
+class ProductsActivity : DataBindingActivity<ActivityProductsBinding>(R.layout.activity_products) {
     private val viewModel: ProductsViewModel by viewModels()
     private val productsAdapter: ProductsAdapter = ProductsAdapter(createAdapterOnClickHandler())
